Translations...
Hi John,
I translated the phrases. In many cases, it is hard to translate because a) i am not familiar with the context b) finnish sentence structure (place of object and subject) is different or reverse than in english. We'll see how it looks like in the site and then it can be improved